use useTerritoryNavigation for sibling territory logic and update local storage keys for separator positions
